rating remade, search added

+++ src/main/java/com/ranv/Model/DTO/RatingDTO.java
@@ -12,7 +12,7 @@ public class RatingDTO {
 
     private Long user;
     private Long manual;
-    private Long value;
+    private int value;
 
 
 }
